support adding key from tree view and focus on key name (#926)
* support adding key from tree view and focus on key name

closes #868
