+++ services/editor/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tweek-editor",
-  "version": "0.4.3",
+  "version": "0.4.4",
   "main": "dist/index.js",
   "repository": "Soluto/tweek",
   "author": "Soluto",
