+++ services/editor/src/pages/keys/components/KeyPage/KeyAddPage/NewKeyInput.js
@@ -49,6 +49,7 @@ const NewKeyInput = compose(
       >
         <ValidationIcon show={isShowingHint} hint={hint} />
         <ComboBox
+          autofocus
           data-field="new-key-name-input"
           suggestions={suggestions}
           value={displayName}
