+++ services/editor/src/pages/keys/components/KeysList/KeysList.css
@@ -109,58 +109,6 @@
   background-color: #3a3e41;
   border-radius: 6px;
 }
-.key-folder {
-  width: 100%;
-}
-.key-folder .key-folder-name {
-  font-size: 16px;
-  display: flex;
-  flex-grow: 1;
-  height: 30px;
-  padding-left: 5px;
-  padding-top: 13px;
-  color: #d8d8d8;
-  background-color: transparent;
-  transition: all 0.2s ease-in-out;
-}
-.key-folder .key-folder-name:hover {
-  cursor: pointer;
-  background-color: #363a3e;
-}
-.key-folder .key-folder-name:active {
-  background-color: #363a3e;
-}
-.key-folder .number-of-folder-keys {
-  font-size: 11px;
-  display: flex;
-  margin-left: 5px;
-  margin-top: 5px;
-  opacity: 0.7;
-}
-.key-folder .number-of-folder-keys:hover {
-  cursor: pointer;
-}
-.key-folder .key-folder-icon {
-  width: 16px;
-  margin-right: 5px;
-  margin-bottom: 10px;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  -webkit-user-select: none;
-  user-select: none;
-  display: inline-block;
-}
-.key-folder .key-folder-icon:hover {
-  cursor: pointer;
-}
-.key-folder .folder-items {
-  display: flex;
-  flex-grow: 1;
-}
-.key-folder .sub-tree {
-  display: flex;
-  flex-grow: 1;
-}
 .search-results {
   width: 100%;
 }
