+++ services/editor/src/pages/keys/components/KeysList/TreeView/TreeView.less
@@ -1,0 +1,64 @@
+@import (reference) "../../../../../styles/core/core.less";
+
+.key-folder {
+    width: 100%;
+    .key-folder-name {
+      font-size: 16px;
+      display: flex;
+      flex-grow: 1;
+      height: 30px;
+      padding-left: 5px;
+      padding-top: 13px;
+      color: #d8d8d8;
+      background-color: transparent;
+      .transitioned();
+      &:hover {
+        cursor: pointer;
+        background-color: #363a3e;
+  
+        .add-key {
+          display: inline-block;
+        }
+      }
+      &:active {
+        background-color: #363a3e;
+      }
+    }
+    .number-of-folder-keys {
+      font-size: 11px;
+      display: flex;
+      margin-left: 5px;
+      margin-top: 5px;
+      opacity: 0.7;
+      &:hover {
+        cursor: pointer;
+      }
+    }
+    .key-folder-icon {
+      width: 16px;
+      margin-right: 5px;
+      margin-bottom: 10px;
+      .user-select(none);
+      display: inline-block;
+      &:hover {
+        cursor: pointer;
+      }
+    }
+    .folder-items {
+      display: flex;
+      flex-grow: 1;
+    }
+    .sub-tree {
+      display: flex;
+      flex-grow: 1;
+    }
+    .add-key {
+      .add-circle-button;
+      display: none;
+      height: 20px;
+      margin-left: auto;
+      margin-right: 10px;
+      cursor: pointer;
+    }
+  }
+
