+++ services/editor/src/store/ducks/selectedKey.js
@@ -95,13 +95,16 @@ const confirmAddKeyAlert = {
   message: 'Adding new key will discard all your changes.\nDo you want to continue?',
 };
 
-export const addKey = shouldShowConfirmationScreen =>
+export const addKey = (shouldShowConfirmationScreen, keyPath) =>
   continueGuard(shouldShowConfirmationScreen, confirmAddKeyAlert, (dispatch) => {
     // update the state to empty key in order to skip on leave hook
     dispatch({ type: KEY_OPENED, payload: createBlankJPadKey() });
     // navigate and set defaults
     dispatch(push('/keys/_blank'));
     dispatch(changeKeyValueType('string'));
+
+    const validation = { isValid: false, hint: '', isShowingHint: false };
+    setImmediate(() => dispatch(updateKeyPath(keyPath, validation)));
   });
 
 export function addKeyDetails() {
